jaseci_kit/jaseci_kit/modules/encoders/requirements.txt
transformers
sentence-transformers
pytorch
fastapi
uvicorn